requirements.txt
torch>=1.12
torchvision>=0.13
torchaudio>=0.12
scikit-learn>=1.0
pandas>=1.3
opencv-python>=4.5
einops>=0.6
Pillow>=9.0
einops>=0.6
tqdm>=4.64
matplotlib>=3.4
seaborn>=0.11
json5>=0.9
pyyaml>=6.0